feat(pci-cold-archive): change routing to data router
handle boundary Error
refactor PageTracking
ref: DTCORE-3104
