+++ packages/manager/apps/pci-cold-archive/src/App.tsx
@@ -1,17 +1,21 @@
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { HashRouter } from 'react-router-dom';
-import RoutesComponent from '@/routes';
+import {
+  createHashRouter,
+  createRoutesFromElements,
+  RouterProvider,
+} from 'react-router-dom';
 import queryClient from './queryClient';
 
+import Routes from '@/routes';
 import '@ovhcloud/ods-themes/default';
+
+const routes = createHashRouter(createRoutesFromElements(Routes));
 
 function App() {
   return (
     <QueryClientProvider client={queryClient}>
-      <HashRouter>
-        <RoutesComponent />
-      </HashRouter>
+      <RouterProvider router={routes} />
       <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
   );
